Added MongoDB support and Children URLs
MongoDB is populated with Chrome History URLs, the HTML in text format for that page, and the links within that page (Child pages). I also cleaned up the links so that we get less garbage links. The DB will not have repeated entries. For now I left the updateDB function to run at startup. Once I finish updating the db directly from chrome extension, we can get rid of this.